Rename components for consistency, deleted commented code, reorder imports
